commit-2

+++ src/main/java/com/yht/shiro2/project/user/entity/User.java
@@ -9,7 +9,7 @@ public class User {
 
     private static final long serialVersionUID = 1L;
     /** 用户ID */
-    private Long userId;
+    private Integer userId;
     /** 部门ID */
     private Long deptId;
     /** 部门父ID */
@@ -47,12 +47,12 @@ public class User {
     /** 岗位组 */
     private Long[] postIds;
 
-    public Long getUserId()
+    public Integer getUserId()
     {
         return userId;
     }
 
-    public void setUserId(Long userId)
+    public void setUserId(Integer userId)
     {
         this.userId = userId;
     }
